Dropped use of narg argument by process_repeated_flag function
It was always going to be problematic, especially given the existence
of repeatable arguments of variable length. (`gdal_translate`, for
instance, takes a `-gcp` argument, typically repeated, which can have
length 4 or 5.) There seems, on the other hand, to be no real downside
to asking that users supply repeated arguments in the form of matrices
in which each row represents an independent supplied argument.