Added tests for equivalency of gdalUtilities and gdalUtils output
Some of the tests fail on my laptop, even though both packages use
GDAL binaries with the same version. This is most likely due to
differences in versions of some of the supporting libraries. (For a
list of 45+ of those supporting libraries, see the example
Makevars.win file displayed at https://github.com/rwinlib/gdal2.