Added gRasterize() plus changes suggested by Joseph Stachelek
On 2018-12-12, Joseph Stachelek submitted a pull request with a number
of changes needed to get the package passing the tests run by
`devtools::check()`. (Thanks for that!) Since my local repo was actually
a couple of commits ahead of the GitHub remote, I did not simply
accept the pull request. Instead, this commit incorporates (most of)
his suggestions, along with some other uncommitted work of my own.

Most notably, it adds a convenience function, `gRasterize()`, that
takes R spatial objects (specifically a `SpatialPointsDataFrame` and a
`RasterLayer` object (from packages **sp** and **raster**,
respectively)) and uses `gdal_rasterize()` to create a new
`RasterLayer` object from the two. I use this function as a drop-in
replacement for the much slower (but still somewhat more general)
`raster::rasterize()`. Will need much more work  to make this
generally useful to others.