More changes suggested by running `R CMD check --as-cran`
Note that, for some reason, the `devtools::check()` convenience
function is failing with an error. Running instead the combination of
`R CMD build` and then `R CMD check --as-cran` gave me a number of
errors, warnings, and notes. At present, I've handled all but two of
the NOTES and one of the WARNings.